Добавлены методы расчета площади

+++ src/Sem03/Figure.java
@@ -1,25 +1,25 @@
 package Sem03;
 
 public abstract class Figure implements Area{
-    int a;
-    int b;
-    int c;
-    int d;
+    double a;
+    double b;
+    double c;
+    double d;
 
-    public Figure(int a, int b, int c, int d) {
+    public Figure(double a, double b, double c, double d) {
         this.a = a;
         this.b = b;
         this.c = c;
         this.d = d;
     }
 
-    public Figure(int a, int b, int c) {
+    public Figure(double a, double b, double c) {
         this.a = a;
         this.b = b;
         this.c = c;
     }
 
-    public Figure(int a) {
+    public Figure(double a) {
         this.a = a;
     }
 }
