+++ src/Sem03/Square.java
@@ -2,7 +2,7 @@ package Sem03;
 
 public class Square extends Figure {
 
-    public Square(int a, int b, int c, int d) {
+    public Square(double a, double b, double c, double d) {
         super(a, b, c, d);
     }
